LayoutTests/fast/scrolling/mac/scrolling-triggerered-layouts-expected.txt


Test that scrolling doesn't trigger a layout per scroll
PASS layoutCount <= 4 is true
PASS successfullyParsed is true

TEST COMPLETE
fixed
